merge

+++ ambari-server/src/main/resources/scripts/bootstrap_agent_env_setup.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-executor=`whoami`
 BINDIR=`dirname "$0"`
 cd $BINDIR
 currentPath=`pwd`
@@ -9,8 +8,16 @@ sshUser=$2
 sshPass=$3
 
 homePath=`cat /etc/passwd | grep tencent: | awk -F':' '{print $6}'`
-
 #copy and execute ambari_agent_env.sh
+if [ "${homePath}" == "" ]
+  then 
+    echo "tencent user has not exists"
+    exit -1
+fi
+if [ ! -f ${homePath}/.ssh/id_rsa.pub ]; then
+  echo "id_rsa.pub does not exists"
+  exit -1; 
+fi
 cp ${homePath}/.ssh/id_rsa.pub ${currentPath}
 tar -czvf ${currentPath}/bootstrap_agent_setup.tar.gz ./id_rsa.pub ./bootstrap_agent_env_script.sh
 ${currentPath}/bootstrap_agent_env_setup.exp ${currentPath}/bootstrap_agent_setup.tar.gz ${hostIP} ${sshUser} ${sshPass}
